requirements.txt
torch>=2.1
transformers>=4.43
sentencepiece>=0.2
spacy>=3.7
nltk>=3.9
textstat>=0.7
scikit-learn>=1.4
streamlit>=1.34
pandas>=2.2
reportlab>=4.0
python-dotenv>=1.0
groq>=0.9

https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
